+++ tests/test_surrogate_vega.sh
@@ -2,13 +2,12 @@
 
 echo $PWD
 
-
-CMD="itwinai exec-pipeline --config config.yaml --pipe-key training_pipeline"
+CMD="itwinai exec-pipeline --config-name config"
 
 SURR_DIR=$PWD/docker/surrogate
 
 if [ ! -d "$SURR_DIR/tmp" ]; then
-  mkdir -p $SURR_DIR/tmp 
+  mkdir -p $SURR_DIR/tmp
 fi
 
 docker build --no-cache -f $SURR_DIR/Dockerfile -t surrogate-test $SURR_DIR
@@ -25,4 +24,4 @@ elif [ "$1" == "gpu" ]; then
 fi
 
 
-docker run -it surrogate-test /bin/bash -c "cd ./use-case && itwinai exec-pipeline --config config.yaml --pipe-key training_pipeline"
+#docker run -it surrogate-test /bin/bash -c "cd ./use-case && itwinai exec-pipeline --config-name config"